refactor: reorganize theme-related components and update imports
