+++ src/front/modules/theme-settings/components/ThemeToggle.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { useTheme } from "../shared/hooks/useTheme";
+import { useTheme } from "../hooks/useTheme";
 import { Sun, Moon } from "lucide-react";
 
 /**
